fix order

+++ src/app/cart/CartController.js
@@ -1,6 +1,5 @@
-const { PrismaClient } = require('@prisma/client');
-const { response } = require('express');
-const { StatusCodes } = require('http-status-codes');
+const { PrismaClient } = require("@prisma/client");
+const { StatusCodes } = require("http-status-codes");
 
 const prisma = new PrismaClient();
 
@@ -8,12 +7,22 @@ const getCarts = async (req, res) => {
   try {
     const response = await prisma.cart.findMany({
       where: { userId: req.user.id },
-      include: {
-        item: true,
-        user: true,
+      select: {
+        id: true,
+        quantity: true,
+        totalprice: true,
+        item: true
       },
     });
-    res.status(StatusCodes.OK).json(response);
+    const totalCartPrice = response.reduce((acc, cart) => {
+      return acc + cart.totalprice;
+    }, 0);
+
+    res.status(StatusCodes.OK).json({
+      status: true,
+      data: response,
+      totalCartPrice,
+    });
   } catch (error) {
     res.status(StatusCodes.INTERNAL_SERVER_ERROR).json({
       msg: error.message,
@@ -22,30 +31,75 @@ const getCarts = async (req, res) => {
 };
 
 const createCarts = async (req, res) => {
-  try {
-    const { userId, itemsId, quantity, totalprice } = req.body;
+  const { itemsId, quantity } = req.body;
+  const userId = req.user.id
 
-    const cart = await prisma.cart.create({
-      data: {
-        userId,
-        itemsId,
-        quantity,
-        totalprice,
+  try {
+    const findCartItem = await prisma.cart.findFirst({
+      where: {
+        itemsId: itemsId
       },
       select: {
+        id: true,
         quantity: true,
-        totalprice: true,
-        item: {
-          select: {
-            price: true,
-            address: true,
-          },
-        },
-      },
+      }
     });
+  const findItem = await prisma.items.findFirst({
+    where: {
+      id: itemsId
+    }
+  })
 
-    res.json(cart).status(StatusCodes.CREATED);
+
+    const sumPrice = quantity * findItem.price
+
+    if (findCartItem) {
+
+        const update = await prisma.cart.update({
+          where: {
+            id: findCartItem.id,
+          },
+          data: {
+            quantity,
+            totalprice: sumPrice
+          },
+          select: {
+            item: {
+              select: {
+                price: true
+              }
+            }
+          }
+        });
+
+        return res.status(StatusCodes.OK).json({
+          success: true,
+          data: update,
+        });
+      
+    }
+      const cart = await prisma.cart.create({
+        data: {
+          userId,
+          itemsId,
+          quantity,
+          totalprice: sumPrice,
+        },
+        select: {
+          id: true,
+          quantity: true,
+          item: {
+            select: {
+              price: true
+            }
+          }
+        }
+      });
+
+      res.status(StatusCodes.CREATED).json(cart);
+    
   } catch (error) {
+    console.log(error)
     res.status(StatusCodes.BAD_REQUEST).json({
       msg: error.message,
     });
@@ -54,7 +108,21 @@ const createCarts = async (req, res) => {
 
 const deleteCarts = async (req, res) => {
   try {
-    const cartId = parseInt(req.params.id);
+    const cartId = req.params.id;
+    console.log(cartId)
+
+    const checkCart = await prisma.cart.findUnique({
+      where:{
+        id: cartId
+      }
+    })
+
+    if(!checkCart){
+      return res.status(StatusCodes.NOT_FOUND).json({
+        success: false,
+        msg: 'Cart Not Found'
+      })
+    }
 
     const cart = await prisma.cart.delete({
       where: {
@@ -62,7 +130,10 @@ const deleteCarts = async (req, res) => {
       },
     });
 
-    res.status(StatusCodes.OK).json(cart);
+    res.status(StatusCodes.OK).json({
+      success: true,
+      msg: 'Success Deleted'
+    });
   } catch (error) {
     res.status(StatusCodes.BAD_REQUEST).json({ msg: error.message });
   }
